+++ requirements.txt
@@ -3,3 +3,4 @@ telegram==0.0.1
 python-telegram-bot==13.11.0
 pytest==7.1.3
 telegram-bot-unittest==0.2
+requests==2.28.1
